+++ packages/builder/builder-shared/src/types/plugin.ts
@@ -59,6 +59,7 @@ export type Plugins = {
   rem: PluginsFn;
   wasm: PluginsFn;
   moment: PluginsFn;
+  externals: PluginsFn;
 };
 
 /**