+++ packages/builder/builder-shared/src/utils.ts
@@ -72,7 +72,9 @@ export function isWebTarget(target: BuilderTarget | BuilderTarget[]) {
   );
 }
 
-export function resolveLoader(loader: string) {
+export function resolvePackage(loader: string, dirname: string) {
   // Vitest do not support require.resolve to source file
-  return process.env.VITEST ? loader : require.resolve(loader);
+  return process.env.VITEST
+    ? loader
+    : require.resolve(loader, { paths: [dirname] });
 }