+++ packages/runtime/plugin-runtime/src/ssr/serverRender/types.ts
@@ -8,6 +8,7 @@ export { RuntimeContext, RenderLevel };
 
 export type SSRPluginConfig = {
   crossorigin?: boolean | 'anonymous' | 'use-credentials';
+  scriptLoading?: 'defer' | 'blocking' | 'module';
   chunkLoadingGlobal?: string;
 } & Exclude<ServerUserConfig['ssr'], boolean>;