+++ packages/builder/builder-cli/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @modern-js/builder-cli
 
+## 2.23.1
+
+### Patch Changes
+
+- Updated dependencies [20c85bb]
+- Updated dependencies [f08bbfc]
+- Updated dependencies [a6b313a]
+- Updated dependencies [5772927]
+- Updated dependencies [811ccd4]
+- Updated dependencies [5a3eeff]
+- Updated dependencies [4af25d9]
+- Updated dependencies [4d4dca0]
+- Updated dependencies [8f2cab0]
+  - @modern-js/builder-webpack-provider@2.23.1
+  - @modern-js/builder-rspack-provider@2.23.1
+  - @modern-js/builder@2.23.1
+  - @modern-js/utils@2.23.1
+  - @modern-js/builder-shared@2.23.1
+  - @modern-js/node-bundle-require@2.23.1
+
 ## 2.23.0
 
 ### Patch Changes