+++ packages/generator/generator-cases/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/generator-cases
 
+## 3.1.23
+
+### Patch Changes
+
+- @modern-js/generator-common@3.1.23
+
 ## 3.1.22
 
 ### Patch Changes