+++ packages/generator/generators/entry-generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/entry-generator
 
+## 3.1.23
+
 ## 3.1.22
 
 ### Patch Changes