+++ packages/generator/generators/monorepo-generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/monorepo-generator
 
+## 3.1.23
+
 ## 3.1.22
 
 ### Patch Changes