+++ packages/runtime/plugin-router-v5/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @modern-js/plugin-router-v5
 
+## 2.23.1
+
+### Patch Changes
+
+- Updated dependencies [f08bbfc]
+- Updated dependencies [a6b313a]
+- Updated dependencies [8f2cab0]
+  - @modern-js/utils@2.23.1
+  - @modern-js/plugin@2.23.1
+  - @modern-js/types@2.23.1
+
 ## 2.23.0
 
 ### Patch Changes