+++ packages/toolkit/compiler/babel/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/babel-compiler
 
+## 2.23.1
+
+### Patch Changes
+
+- Updated dependencies [f08bbfc]
+- Updated dependencies [a6b313a]
+- Updated dependencies [8f2cab0]
+  - @modern-js/utils@2.23.1
+
 ## 2.23.0
 
 ### Patch Changes