+++ scripts/prebundle/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @scripts/prebundle
 
+## 2.23.1
+
+### Patch Changes
+
+- @modern-js/tsconfig@2.23.1
+- @scripts/build@2.23.1
+
 ## 2.23.0
 
 ### Patch Changes